refactor: 웹소켓 핸드쉐이크 파일 위치 변경
